pyproject.toml
[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[project]
name = "gasregnet"
version = "0.0.0"
description = "Genome-neighborhood mining for bacterial gas-sensing regulator discovery."
readme = "README.md"
requires-python = "==3.11.9"
license = { text = "MIT" }
authors = [
  { name = "GasRegNet contributors" }
]
dependencies = [
  "polars==1.12.0",
  "pandera[polars]==0.23.1",
  "pydantic==2.9.2",
  "PyYAML==6.0.3",
  "snakemake==8.30.0",
  "duckdb==1.1.3",
  "pyarrow==17.0.0",
  "structlog==24.4.0",
  "typer==0.12.5",
  "scipy==1.13.1",
  "statsmodels==0.14.4",
  "matplotlib==3.9.2",
  "seaborn==0.13.2",
  "biopython==1.84",
  "pyhmmer==0.10.15"
]

[project.optional-dependencies]
dev = [
  "ruff==0.7.4",
  "mypy==1.11.2",
  "pytest==8.3.3",
  "hypothesis==6.115.5",
  "pytest-cov==5.0.0"
]
embeddings = [
  "torch>=2.0",
  "fair-esm>=2.0"
]
bayes = [
  "pymc>=5.0"
]
provenance = [
  "prov>=2.0"
]

[project.scripts]
gasregnet = "gasregnet.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["gasregnet"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["gasregnet"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "scaling: opt-in scaling acceptance probes"
]

[tool.coverage.run]
source = ["gasregnet"]
